# Speed up v3 security groups