@sethboyles 2021-05-05
Thanks for the PR! This all makes sense and I will merge shortly.

Testing out the latter (worst) case on our environment, we saw a slight decrease in response time, from ~2m28s to ~2m9s

FWIW I experimented a little with the latter case, replacing `dataset.select(:guid).all` with `dataset.select_map(:guid)`, the theory being that `select_map` avoids constructing a Ruby object per Space record, and instead simply returns an array, but only saw a minimal decrease in response time (from 2m9s -> 2m4s)

Additionally, I tried using eager loading, which batches all 2000 spaces into one query.  Later in the presenter would have to be changed to use the already loaded array of Space records rather than a `dataset`.  The queries themselves would execute quickly, but the Sequel ORM still takes a long time to stitch the 1 million associations together.  The request ended up being longer than 3 minutes overall.  I *think* eager loading might have had potential for your first case example, but given the improvement already shown I didn't try it out.